requirements.txt
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
PyYAML>=6.0
scipy>=1.13
seaborn>=0.13
torch>=2.2
torchdiffeq>=0.2.4
tqdm>=4.66
pytest>=8.0
